requirements.txt
agno==0.1.7
google-generativeai==0.6.0
streamlit==1.34.0
